refactor: ui/utils/icons 공통으로 분리
