+++ src/basic/entities/index.ts
@@ -1,11 +1,4 @@
-// Product
 export * from './product';
-
-// Cart
 export * from './cart';
-
-// Coupon
 export * from './coupon';
-
-// Payment
 export * from './payment';
