+++ src/basic/entities/payment/components/Payment.tsx
@@ -1,7 +1,7 @@
 import { CartItemType, CouponType } from '@/types';
 import { usePayment } from '../hooks/usePayment';
-import { Button } from '../../../ui';
-import { calculateCartTotal } from '../../../utils';
+import { Button } from '@/ui';
+import { calculateCartTotal } from '@/utils';
 
 interface PaymentPropsType {
   addNotification: (message: string, type: 'success' | 'error') => void;
