+++ src/icons/CartIcon.tsx
