+++ src/icons/TrashIcon.tsx
