+++ src/icons/XIcon.tsx
