+++ src/utils/discountUtils.ts
