+++ src/utils/productUtils.ts
