+++ tsconfig.app.json
@@ -11,7 +11,10 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"],
-      "@/types": ["src/types"]
+      "@/types": ["src/types"],
+      "@/utils": ["src/utils"],
+      "@/icons/*": ["src/icons/*"],
+      "@/ui/*": ["src/ui/*"]
     },
 
     /* Bundler mode */
