+++ vite.config.ts
@@ -1,10 +1,20 @@
 import { defineConfig as defineTestConfig, mergeConfig } from 'vitest/config';
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
+import { fileURLToPath } from 'url';
+import { dirname, resolve } from 'path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 export default mergeConfig(
   defineConfig({
     plugins: [react()],
+    resolve: {
+      alias: {
+        '@': resolve(__dirname, './src'),
+      },
+    },
   }),
   defineTestConfig({
     test: {
